fix(core): widen the numbers OBS does not bound
Found by the request sweep: GetOutputList became unreadable because an
idle virtual camera reported outputHeight as 2586032160. The field is
uint32_t in libobs and obs-websocket passes it through unclamped, so an
int could not hold it, and one bad field fails the whole response rather
than just itself.

Swept the rest of the numeric table against the obs-websocket and
obs-studio sources. A field is safe as an int only where obs-websocket
validates a range; where it copies out of libobs or a settings blob the C
type decides. That reclassifies:

  outputWidth/Height (stub)  uint32_t, unclamped
  render/output frames       uint32_t, monotonic
  webSocketSession messages  uint64_t
  sceneItemId                int64_t, validated >= 0 with no upper bound
  transitionDuration         int64_t out of private settings
  alignment/boundsAlignment  uint32_t, validated to the full range

sceneIndex becomes nullable for the same reason: GetCanvasSceneList has
no index to report and sends null.

Removes FindSceneItemIdInt32Async, which named a return type it no longer
has.

+++ ObsWebSocket.Core/Handles/ObsHandles.cs
@@ -157,7 +157,7 @@ public sealed record SceneHandle : IObsHandle
     /// <summary>
     /// A scene item in this scene, by the numeric id OBS assigned it.
     /// </summary>
-    public SceneItemHandle Item(int sceneItemId) => SceneItemHandle.For(this, sceneItemId);
+    public SceneItemHandle Item(long sceneItemId) => SceneItemHandle.For(this, sceneItemId);
 
     /// <summary>
     /// A scene item in this scene, by the name of the source it shows. Needs resolving, because
@@ -297,7 +297,7 @@ public sealed record SourceHandle : IObsHandle
 /// </remarks>
 public sealed record SceneItemHandle
 {
-    private SceneItemHandle(SceneHandle scene, int sceneItemId)
+    private SceneItemHandle(SceneHandle scene, long sceneItemId)
     {
         Scene = scene;
         SceneItemId = sceneItemId;
@@ -307,10 +307,10 @@ public sealed record SceneItemHandle
     public SceneHandle Scene { get; }
 
     /// <summary>The numeric id OBS assigned the item.</summary>
-    public int SceneItemId { get; }
+    public long SceneItemId { get; }
 
     /// <summary>Builds a handle for an id already known.</summary>
-    public static SceneItemHandle For(SceneHandle scene, int sceneItemId)
+    public static SceneItemHandle For(SceneHandle scene, long sceneItemId)
     {
         ArgumentNullException.ThrowIfNull(scene);
         ArgumentOutOfRangeException.ThrowIfNegative(sceneItemId);
